via browser not GIT
Erase the Internal communicator class because it's not needed. the ID of the Messages class can be linked to the conference Program,
Erase conference class (which contains only title) becase it's already covered in the  conferenceProgram class. ConferenceProgram class is used to handle information about the conference name, the topic of the conference, etc in which will be used in FR2.0 and FR before2.0 (which contains the data input for the conferenceProgram Class)
Erase the session program class because session program only used to see the sessions of the conference program and the detail of the sessions which can be presented by schedule using an operation-> which I added to the class diagram- viewSessionProgram and viewSelectedSession(SessionID).